src/requirements.txt
# global requirements
gunicorn==22.0.0
azure-monitor-opentelemetry

# django requirements
Django==5.1.3
django-import-export
django-json-widget
django-markdownify
jsonref
jsonschema
jsonschema-specifications
Markdown
openai
PyYAML
# martor
djangorestframework
psycopg
setuptools
django-cms
django-filer
djangocms_text_ckeditor
djangocms-frontend
djangocms-versioning
djangocms-alias
djangocms_admin_style